Might fix validation error #538 -
The validation error came from this command on the Travis CI build:

./validate.sh --nofmt --cov --race 10

The command worked for me when I ran it locally on go version 1.10.1 and 1.9.3.
 * go version go1.9.3 darwin/amd64
 * go version go1.10.1 darwin/amd64
 * both installed with homebrew

I then did a pull and ran the command on Beachfront's staging server
go version go1.10 linux/amd64

That reproduced the error :
gofmt needs to be run, 1 files have issues.  Below is a list of files to review:
adapters/beachfront/beachfront.go

So I ran go fmt to see what it would fix. The diff of this commit will show
that the problem was several conditionals where I, being new to Go, used
parenthesis. The Darwin builds were letting that slide for whatever reason.
The Linux builds were not (I assume that Travis CL is running in a Linux VM
somewhere).