Currency support #280 - step 1
This CL allows prebid server to reject any bids having a currency mismatch between allowed
currencies in the bid request and the currency declared in bids.

For the time being, if allowed currencies are not declared in bid request or in bid
the currency is implicitly set to `USD` to prevent from any breaking changes.

This CL includes:
- Prebid server to reject bids having currency mistmatch

Tnis CL doesn't include:
- Changing the default currency (which is USD)