Bid deduplication fix
In case we find bids with same category or pb_dur we randomly deduplicate one of them. This can a bid that was previously processed or bid on current iteration. In case we deduplicate bid from previous iteration and this is the only one bid in bidder - we need to remove the entire bidder from processing. Without this fix we removed "current iteration" bidder instead of "previously processed" . bidder.