Currency: Intermediate rate logic
Closes #3566

The Prebid.js currency module is able to cross convert currencies using USD (or GBP) as an intermediate.
This feature is not currently implemented in PBS Go, so the following error exists:
 'Unable to convert from currency NOK to desired ad server currency SEK'.
In this PR we implement the cross currency conversion logic through
an intermediate currency as a part of the GetRate method in the currency package.